+++ core/src/db/Utils.cpp
@@ -21,6 +21,7 @@
 #include <vector>
 
 #include "cache/CpuCacheMgr.h"
+#include "db/snapshot/Resources.h"
 #ifdef MILVUS_GPU_VERSION
 #include "cache/GpuCacheMgr.h"
 #endif
@@ -250,7 +251,7 @@ GetIndexName(int32_t index_type) {
         {(int32_t)engine::EngineType::FAISS_IDMAP, "IDMAP"},
         {(int32_t)engine::EngineType::FAISS_IVFFLAT, "IVFFLAT"},
         {(int32_t)engine::EngineType::FAISS_IVFSQ8, "IVFSQ8"},
-        {(int32_t)engine::EngineType::NSG_MIX, "NSG"},
+        {(int32_t)engine::EngineType::FAISS_IVFSQ8NR, "IVFSQ8NR"},
         {(int32_t)engine::EngineType::FAISS_IVFSQ8H, "IVFSQ8H"},
         {(int32_t)engine::EngineType::FAISS_PQ, "PQ"},
         {(int32_t)engine::EngineType::SPTAG_KDT, "KDT"},
@@ -258,8 +259,8 @@ GetIndexName(int32_t index_type) {
         {(int32_t)engine::EngineType::FAISS_BIN_IDMAP, "IDMAP"},
         {(int32_t)engine::EngineType::FAISS_BIN_IVFFLAT, "IVFFLAT"},
         {(int32_t)engine::EngineType::HNSW_SQ8NR, "HNSW_SQ8NR"},
-        {(int32_t)engine::EngineType::FAISS_IVFSQ8NR, "FAISS_IVFSQ8NR"},
         {(int32_t)engine::EngineType::HNSW, "HNSW"},
+        {(int32_t)engine::EngineType::NSG_MIX, "NSG"},
         {(int32_t)engine::EngineType::ANNOY, "ANNOY"}};
 
     if (index_type_name.find(index_type) == index_type_name.end()) {
@@ -300,6 +301,42 @@ EraseFromCache(const std::string& item_key) {
         cache::GpuCacheMgr::GetInstance(gpu)->EraseItem(item_key);
     }
 #endif
+}
+
+Status
+CreatePath(const snapshot::Segment* segment, const DBOptions& options, std::string& path) {
+    std::string tables_path = options.meta_.path_ + TABLES_FOLDER;
+    STATUS_CHECK(CommonUtil::CreateDirectory(tables_path));
+    std::string collection_path = tables_path + "/" + std::to_string(segment->GetCollectionId());
+    STATUS_CHECK(CommonUtil::CreateDirectory(collection_path));
+    std::string partition_path = collection_path + "/" + std::to_string(segment->GetPartitionId());
+    STATUS_CHECK(CommonUtil::CreateDirectory(partition_path));
+    path = partition_path + "/" + std::to_string(segment->GetID());
+    STATUS_CHECK(CommonUtil::CreateDirectory(path));
+
+    return Status::OK();
+}
+
+Status
+CreatePath(const snapshot::Partition* partition, const DBOptions& options, std::string& path) {
+    std::string tables_path = options.meta_.path_ + TABLES_FOLDER;
+    STATUS_CHECK(CommonUtil::CreateDirectory(tables_path));
+    std::string collection_path = tables_path + "/" + std::to_string(partition->GetCollectionId());
+    STATUS_CHECK(CommonUtil::CreateDirectory(collection_path));
+    path = collection_path + "/" + std::to_string(partition->GetID());
+    STATUS_CHECK(CommonUtil::CreateDirectory(path));
+
+    return Status::OK();
+}
+
+Status
+CreatePath(const snapshot::Collection* collection, const DBOptions& options, std::string& path) {
+    std::string tables_path = options.meta_.path_ + TABLES_FOLDER;
+    STATUS_CHECK(CommonUtil::CreateDirectory(tables_path));
+    path = tables_path + "/" + std::to_string(collection->GetID());
+    STATUS_CHECK(CommonUtil::CreateDirectory(path));
+
+    return Status::OK();
 }
 
 }  // namespace utils