+++ core/src/index/knowhere/knowhere/index/vector_index/IndexIVFSQ.cpp
@@ -16,6 +16,8 @@
 #include <faiss/gpu/GpuAutoTune.h>
 #include <faiss/gpu/GpuCloner.h>
 #endif
+#include <faiss/IndexFlat.h>
+#include <faiss/IndexScalarQuantizer.h>
 #include <faiss/clone_index.h>
 #include <faiss/index_factory.h>
 
@@ -35,14 +37,18 @@ void
 IVFSQ::Train(const DatasetPtr& dataset_ptr, const Config& config) {
     GETTENSOR(dataset_ptr)
 
-    std::stringstream index_type;
-    index_type << "IVF" << config[IndexParams::nlist] << ","
-               << "SQ" << config[IndexParams::nbits];
-    auto build_index =
-        faiss::index_factory(dim, index_type.str().c_str(), GetMetricType(config[Metric::TYPE].get<std::string>()));
-    build_index->train(rows, (float*)p_data);
+    // std::stringstream index_type;
+    // index_type << "IVF" << config[IndexParams::nlist] << ","
+    //           << "SQ" << config[IndexParams::nbits];
+    // index_ = std::shared_ptr<faiss::Index>(
+    //    faiss::index_factory(dim, index_type.str().c_str(), GetMetricType(config[Metric::TYPE].get<std::string>())));
 
-    index_.reset(faiss::clone_index(build_index));
+    faiss::MetricType metric_type = GetMetricType(config[Metric::TYPE].get<std::string>());
+    faiss::Index* coarse_quantizer = new faiss::IndexFlat(dim, metric_type);
+    index_ = std::shared_ptr<faiss::Index>(new faiss::IndexIVFScalarQuantizer(
+        coarse_quantizer, dim, config[IndexParams::nlist].get<int64_t>(), faiss::QuantizerType::QT_8bit, metric_type));
+
+    index_->train(rows, (float*)p_data);
 }
 
 VecIndexPtr