+++ core/src/index/knowhere/knowhere/index/vector_index/IndexIVFSQNR.cpp
@@ -9,68 +9,60 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License
 
+#include <memory>
 #include <string>
 
-#include <faiss/IndexFlat.h>
-#include <faiss/IndexIVFPQ.h>
-#include <faiss/clone_index.h>
 #ifdef MILVUS_GPU_VERSION
+#include <faiss/gpu/GpuAutoTune.h>
 #include <faiss/gpu/GpuCloner.h>
 #endif
+#include <faiss/IndexFlat.h>
+#include <faiss/IndexScalarQuantizer.h>
+#include <faiss/clone_index.h>
+#include <faiss/index_factory.h>
 
 #include "knowhere/common/Exception.h"
-#include "knowhere/common/Log.h"
-#include "knowhere/index/vector_index/IndexIVFPQ.h"
+#include "knowhere/index/vector_index/IndexIVFSQNR.h"
 #include "knowhere/index/vector_index/adapter/VectorAdapter.h"
 #include "knowhere/index/vector_index/helpers/IndexParameter.h"
 #ifdef MILVUS_GPU_VERSION
-#include "knowhere/index/vector_index/gpu/IndexGPUIVF.h"
-#include "knowhere/index/vector_index/gpu/IndexGPUIVFPQ.h"
+#include "knowhere/index/vector_index/gpu/IndexGPUIVFSQNR.h"
+#include "knowhere/index/vector_index/helpers/FaissGpuResourceMgr.h"
 #endif
 
 namespace milvus {
 namespace knowhere {
 
 void
-IVFPQ::Train(const DatasetPtr& dataset_ptr, const Config& config) {
+IVFSQNR::Train(const DatasetPtr& dataset_ptr, const Config& config) {
     GETTENSOR(dataset_ptr)
 
     faiss::MetricType metric_type = GetMetricType(config[Metric::TYPE].get<std::string>());
     faiss::Index* coarse_quantizer = new faiss::IndexFlat(dim, metric_type);
-    index_ = std::shared_ptr<faiss::Index>(new faiss::IndexIVFPQ(
-        coarse_quantizer, dim, config[IndexParams::nlist].get<int64_t>(), config[IndexParams::m].get<int64_t>(),
-        config[IndexParams::nbits].get<int64_t>(), metric_type));
+    index_ = std::shared_ptr<faiss::Index>(
+        new faiss::IndexIVFScalarQuantizer(coarse_quantizer, dim, config[IndexParams::nlist].get<int64_t>(),
+                                           faiss::QuantizerType::QT_8bit, metric_type, false));
 
     index_->train(rows, (float*)p_data);
 }
 
 VecIndexPtr
-IVFPQ::CopyCpuToGpu(const int64_t device_id, const Config& config) {
+IVFSQNR::CopyCpuToGpu(const int64_t device_id, const Config& config) {
 #ifdef MILVUS_GPU_VERSION
     if (auto res = FaissGpuResourceMgr::GetInstance().GetRes(device_id)) {
         ResScope rs(res, device_id, false);
+
         auto gpu_index = faiss::gpu::index_cpu_to_gpu(res->faiss_res.get(), device_id, index_.get());
 
         std::shared_ptr<faiss::Index> device_index;
         device_index.reset(gpu_index);
-        return std::make_shared<GPUIVFPQ>(device_index, device_id, res);
+        return std::make_shared<GPUIVFSQNR>(device_index, device_id, res);
     } else {
         KNOWHERE_THROW_MSG("CopyCpuToGpu Error, can't get gpu_resource");
     }
 #else
-    KNOWHERE_THROW_MSG("Calling IVFPQ::CopyCpuToGpu when we are using CPU version");
+    KNOWHERE_THROW_MSG("Calling IVFSQ::CopyCpuToGpu when we are using CPU version");
 #endif
-}
-
-std::shared_ptr<faiss::IVFSearchParameters>
-IVFPQ::GenParams(const Config& config) {
-    auto params = std::make_shared<faiss::IVFPQSearchParameters>();
-    params->nprobe = config[IndexParams::nprobe];
-    // params->scan_table_threshold = config["scan_table_threhold"]
-    // params->polysemous_ht = config["polysemous_ht"]
-    // params->max_codes = config["max_codes"]
-
-    return params;
 }
 
 }  // namespace knowhere