+++ core/src/index/unittest/Helper.h
@@ -15,6 +15,7 @@
 #include "knowhere/index/vector_index/IndexIVF.h"
 #include "knowhere/index/vector_index/IndexIVFPQ.h"
 #include "knowhere/index/vector_index/IndexIVFSQ.h"
+#include "knowhere/index/vector_index/IndexIVFSQNR.h"
 #include "knowhere/index/vector_index/IndexType.h"
 #include "knowhere/index/vector_index/helpers/IndexParameter.h"
 #include "knowhere/index/vector_offset_index/IndexIVF_NM.h"
@@ -23,6 +24,7 @@
 #include "knowhere/index/vector_index/gpu/IndexGPUIVF.h"
 #include "knowhere/index/vector_index/gpu/IndexGPUIVFPQ.h"
 #include "knowhere/index/vector_index/gpu/IndexGPUIVFSQ.h"
+#include "knowhere/index/vector_index/gpu/IndexGPUIVFSQNR.h"
 #include "knowhere/index/vector_index/gpu/IndexIVFSQHybrid.h"
 #include "knowhere/index/vector_offset_index/gpu/IndexGPUIVF_NM.h"
 #endif
@@ -45,6 +47,8 @@ IndexFactory(const milvus::knowhere::IndexType& type, const milvus::knowhere::In
             return std::make_shared<milvus::knowhere::IVFPQ>();
         } else if (type == milvus::knowhere::IndexEnum::INDEX_FAISS_IVFSQ8) {
             return std::make_shared<milvus::knowhere::IVFSQ>();
+        } else if (type == milvus::knowhere::IndexEnum::INDEX_FAISS_IVFSQ8NR) {
+            return std::make_shared<milvus::knowhere::IVFSQNR>();
         } else if (type == milvus::knowhere::IndexEnum::INDEX_FAISS_IVFSQ8H) {
             std::cout << "IVFSQ8H does not support MODE_CPU" << std::endl;
         } else {
@@ -58,6 +62,8 @@ IndexFactory(const milvus::knowhere::IndexType& type, const milvus::knowhere::In
             return std::make_shared<milvus::knowhere::GPUIVFPQ>(DEVICEID);
         } else if (type == milvus::knowhere::IndexEnum::INDEX_FAISS_IVFSQ8) {
             return std::make_shared<milvus::knowhere::GPUIVFSQ>(DEVICEID);
+        } else if (type == milvus::knowhere::IndexEnum::INDEX_FAISS_IVFSQ8NR) {
+            return std::make_shared<milvus::knowhere::GPUIVFSQNR>(DEVICEID);
         } else if (type == milvus::knowhere::IndexEnum::INDEX_FAISS_IVFSQ8H) {
             return std::make_shared<milvus::knowhere::IVFSQHybrid>(DEVICEID);
         } else {
@@ -111,6 +117,7 @@ class ParamGenerator {
                 {milvus::knowhere::meta::DEVICEID, DEVICEID},
             };
         } else if (type == milvus::knowhere::IndexEnum::INDEX_FAISS_IVFSQ8 ||
+                   type == milvus::knowhere::IndexEnum::INDEX_FAISS_IVFSQ8NR ||
                    type == milvus::knowhere::IndexEnum::INDEX_FAISS_IVFSQ8H) {
             return milvus::knowhere::Config{
                 {milvus::knowhere::meta::DIM, DIM},